continued renaming MVVM project into FindMyMeasure.Gui
