+++ FindMyMeasure.Gui/App.config
@@ -2,7 +2,7 @@
 <configuration>
     <configSections>
         <sectionGroup name="userSettings" type="System.Configuration.UserSettingsGroup, System, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089">
-            <section name="FindMyMeasure.Gui.MVVM.Properties.Settings" type="System.Configuration.ClientSettingsSection, System, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" allowExeDefinition="MachineToLocalUser" requirePermission="false" />
+            <section name="FindMyMeasure.Gui.Properties.Settings" type="System.Configuration.ClientSettingsSection, System, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" allowExeDefinition="MachineToLocalUser" requirePermission="false" />
         </sectionGroup>
     </configSections>
     <startup> 
@@ -38,7 +38,7 @@
     </assemblyBinding>
   </runtime>
   <userSettings>
-    <FindMyMeasure.Gui.MVVM.Properties.Settings>
+    <FindMyMeasure.Gui.Properties.Settings>
       <setting name="AnalyseHiddenPages" serializeAs="String">
         <value>True</value>
       </setting>
@@ -48,6 +48,6 @@
       <setting name="LastRunSavePath" serializeAs="String">
         <value>lastRun.json</value>
       </setting>
-    </FindMyMeasure.Gui.MVVM.Properties.Settings>
+    </FindMyMeasure.Gui.Properties.Settings>
   </userSettings>
 </configuration>
