+++ FindMyMeasure.Gui/App.xaml
@@ -1,7 +1,7 @@
-﻿<Application x:Class="FindMyMeasure.Gui.MVVM.App"
+﻿<Application x:Class="FindMyMeasure.Gui.App"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:local="clr-namespace:FindMyMeasure.Gui.MVVM"
+             xmlns:local="clr-namespace:FindMyMeasure.Gui"
              StartupUri="ReportSelectionWindow.xaml">
     <Application.Resources>
         <ResourceDictionary Source="Resources/StringResources.xaml"/>
