+++ FindMyMeasure.Gui/App.xaml.cs
@@ -6,7 +6,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using System.Windows;
 
-namespace FindMyMeasure.Gui.MVVM
+namespace FindMyMeasure.Gui
 {
     /// <summary>
     /// Logique d'interaction pour App.xaml
