+++ FindMyMeasure.Gui/FindMyMeasure.Gui.csproj
@@ -6,8 +6,8 @@
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
     <ProjectGuid>{4C1F3996-4CA5-4F4A-8C68-AB836FE307F6}</ProjectGuid>
     <OutputType>WinExe</OutputType>
-    <RootNamespace>FindMyMeasure.Gui.MVVM</RootNamespace>
-    <AssemblyName>FindMyMeasure.Gui.MVVM</AssemblyName>
+    <RootNamespace>FindMyMeasure.Gui</RootNamespace>
+    <AssemblyName>FindMyMeasure.Gui</AssemblyName>
     <TargetFrameworkVersion>v4.7.2</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <ProjectTypeGuids>{60dc8134-eba5-43b8-bcc9-bb4bc16c2548};{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}</ProjectTypeGuids>
