+++ FindMyMeasure.Gui/LoadingProgressWindow.xaml
@@ -1,9 +1,9 @@
-﻿<Window x:Class="FindMyMeasure.Gui.MVVM.LoadingProgressWindow"
+﻿<Window x:Class="FindMyMeasure.Gui.LoadingProgressWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:FindMyMeasure.Gui.MVVM"
+        xmlns:local="clr-namespace:FindMyMeasure.Gui"
         mc:Ignorable="d"
         Title="{DynamicResource LoadingProgressWindow.Title}" Height="300" Width="500">
     <Grid>
