+++ FindMyMeasure.Gui/Models/ReportAnalysisResult.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace FindMyMeasure.Gui.MVVM.Models
+namespace FindMyMeasure.Gui.Models
 {
     public class ReportAnalysisResult
     {
