+++ FindMyMeasure.Gui/Properties/AssemblyInfo.cs
@@ -7,11 +7,11 @@ using System.Windows;
 // Les informations générales relatives à un assembly dépendent de
 // l'ensemble d'attributs suivant. Pour modifier les informations
 // associées à un assembly.
-[assembly: AssemblyTitle("FindMyMeasure.Gui.MVVM")]
+[assembly: AssemblyTitle("FindMyMeasure.Gui")]
 [assembly: AssemblyDescription("")]
 [assembly: AssemblyConfiguration("")]
 [assembly: AssemblyCompany("HP")]
-[assembly: AssemblyProduct("FindMyMeasure.Gui.MVVM")]
+[assembly: AssemblyProduct("FindMyMeasure.Gui")]
 [assembly: AssemblyCopyright("Copyright © HP 2025")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
