+++ FindMyMeasure.Gui/ReportSelectionWindow.xaml
@@ -1,10 +1,10 @@
 ﻿
-<Window x:Class="FindMyMeasure.Gui.MVVM.ReportSelectionWindow"
+<Window x:Class="FindMyMeasure.Gui.ReportSelectionWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:FindMyMeasure.Gui.MVVM"
+        xmlns:local="clr-namespace:FindMyMeasure.Gui"
         mc:Ignorable="d"
         Title="{DynamicResource ReportSelectionWindow.Title}" Height="450" Width="800">
