+++ FindMyMeasure.Gui/ReportSelectionWindow.xaml.cs
@@ -1,8 +1,8 @@
 ﻿using FindMyMeasure.Database;
 using FindMyMeasure.Enums;
-using FindMyMeasure.Gui.MVVM.Exceptions;
-using FindMyMeasure.Gui.MVVM.Models;
-using FindMyMeasure.Gui.MVVM.ViewModels;
+using FindMyMeasure.Gui.Exceptions;
+using FindMyMeasure.Gui.Models;
+using FindMyMeasure.Gui.ViewModels;
 using FindMyMeasure.PowerBI;
 using Microsoft.Win32;
 using System;
@@ -16,10 +16,10 @@ using System.Text.Json.Nodes;
 using System.Threading.Tasks;
 using System.Windows;
 using System.Windows.Controls;
-using static FindMyMeasure.Gui.MVVM.Models.ReportAnalysisConfiguration;
+using static FindMyMeasure.Gui.Models.ReportAnalysisConfiguration;
 
 
-namespace FindMyMeasure.Gui.MVVM
+namespace FindMyMeasure.Gui
 {
     /// <summary>
     /// Logique d'interaction pour MainWindow.xaml
